Favourite, search, review functionality & overall integration testing

+++ Backend/Data/ReviewRepository.cs
@@ -1,4 +1,5 @@
 using Backend.Data.Abstraction;
+using Backend.DTOs;
 using Backend.Models;
 using Backend.Utils;
 using Microsoft.EntityFrameworkCore;
@@ -12,24 +13,20 @@ internal class ReviewRepository : IReviewRepository
         _context = context;
     }
 
-        public async Task<Review> GetReview(int userId, long movieId)
+    public async Task<Review?> GetReview(int userId, long movieId)
     {
-        var review = await _context.Reviews.FirstOrDefaultAsync(x => x.UserId == userId && x.MovieId == movieId);
-        if (review == null)
-        {
-            return null;
-        }
-        return review;
+        return await _context.Reviews.FirstOrDefaultAsync(x => x.UserId == userId && x.MovieId == movieId) ?? null;
     }
+
     public async Task SetReview(Review review)
     {
-
         var rev = await _context.Reviews.FirstOrDefaultAsync(x => x.UserId == review.UserId && x.MovieId == review.MovieId);
         if (rev != null)
         {
             rev.ReviewText = review.ReviewText;
-             _context.Entry(rev).State = EntityState.Modified;
-        } else
+            _context.Entry(rev).State = EntityState.Modified;
+        }
+        else
         {
             await _context.Reviews.AddAsync(review);
         }
@@ -39,13 +36,18 @@ internal class ReviewRepository : IReviewRepository
     public async Task DeleteReview(int id)
     {
         var review = await _context.Reviews.FirstOrDefaultAsync(x => x.Id == id);
-         _context.Reviews.Remove(review);
+        if (review == null)
+        {
+            return;
+        }
+        _context.Reviews.Remove(review);
+        _context.Entry(review).State = EntityState.Deleted;
         await _context.SaveChangesAsync();
     }
 
     public async Task<IList<Review>> GetMovieReviews(long movieId)
     {
-        var reviews =  await _context.Reviews.Include(x=>x.User).Where(x=>x.MovieId == movieId).ToListAsync();
+        var reviews = await _context.Reviews.Include(x => x.User).Where(x => x.MovieId == movieId).ToListAsync();
         return reviews;
     }
 }
